requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langgraph
openai
streamlit
python-dotenv
faiss-cpu
pypdf
python-pptx
fastapi
uvicorn[standard]
python-multipart
requests
docx2txt
pydantic
typing-extensions